+++ core/src/server/web_impl/Constants.cpp
@@ -9,8 +9,8 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "knowhere/index/IndexType.h"
 #include "server/web_impl/Constants.h"
+#include "knowhere/index/IndexType.h"
 
 namespace milvus {
 namespace server {